// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FeedbackType = {
  BUG: 'BUG',
  FEATURE: 'FEATURE',
  OTHER: 'OTHER'
} as const

export type FeedbackType = (typeof FeedbackType)[keyof typeof FeedbackType]


export const Sentiment = {
  POSITIVE: 'POSITIVE',
  NEUTRAL: 'NEUTRAL',
  NEGATIVE: 'NEGATIVE'
} as const

export type Sentiment = (typeof Sentiment)[keyof typeof Sentiment]